text-semantic-search/semantic_search/requirements.txt
pip==19.0.2
httplib2==0.12.1
oauth2client==4.1.3
google-api-python-client==1.7.8
google-cloud-storage==1.13.2
google-cloud-datastore==1.7.3
tensorflow==1.12
tensorflow-hub==0.2
annoy==1.15.0
Flask==1.0.2
gunicorn==19.9.0